fazendo autorizacao do backend, ajustando rotas e melhorando tratamento de autorizacao error 403

+++ projeto/backend/controllers/itemCompraController.js
@@ -1,46 +1,63 @@
 const ItemCompraService = require("../services/ItemCompraService");
+const { AutorizacaoService, ROLES } = require("../services/AutorizacaoService");
 
 exports.get = async (req, res) => {
-  let id = req.params.id;
+  if (AutorizacaoService.validarRoles(req, [ROLES.ESTOQUE, ROLES.MASTER])) {
+    let id = req.params.id;
 
-  try {
-    const item = await ItemCompraService.getItemComprabyId(id);
-    res.json(item);
-  } catch (error) {
-    res.status(500).json({ error: error.message });
+    try {
+      const item = await ItemCompraService.getItemComprabyId(id);
+      res.json(item);
+    } catch (error) {
+      res.status(500).json({ error: error.message });
+    }
+  } else {
+    res.status(403).json({ error: 'Acesso negado' });
   }
 };
 
 exports.getAll = async (req, res) => {
-  try {
-    const registros = await ItemCompraService.getAllItensCompras();
+  if (AutorizacaoService.validarRoles(req, [ROLES.ESTOQUE, ROLES.MASTER])) {
+    try {
+      const registros = await ItemCompraService.getAllItensCompras();
 
-    if (!registros) {
-      return res.status(404).json("Não existem itens de compras cadastradas!");
+      if (!registros) {
+        return res.status(404).json("Não existem itens de compras cadastradas!");
+      }
+
+      res.json(registros);
+    } catch (err) {
+      return res.status(500).json({ error: err.message });
     }
-
-    res.json(registros);
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
+  } else {
+    res.status(403).json({ error: 'Acesso negado' });
   }
 };
 
 exports.add = async (req, res) => {
-  try {
-    const registro = await ItemCompraService.addItemCompra(req.body);
-    res.status(201).json(registro);
-  } catch (error) {
-    res.status(500).json({ error: error.message });
+  if (AutorizacaoService.validarRoles(req, [ROLES.ESTOQUE, ROLES.MASTER])) {
+    try {
+      const registro = await ItemCompraService.addItemCompra(req.body);
+      res.status(201).json(registro);
+    } catch (error) {
+      res.status(500).json({ error: error.message });
+    }
+  } else {
+    res.status(403).json({ error: 'Acesso negado' });
   }
 };
 
 exports.delete = async (req, res) => {
-  let id = req.params.id;
+  if (AutorizacaoService.validarRoles(req, [ROLES.ESTOQUE, ROLES.MASTER])) {
+    let id = req.params.id;
 
-  try {
-    const registro = await ItemCompraService.deleteItemCompra(id);
-    res.json(registro);
-  } catch (error) {
-    res.status(500).json({ error: error.message });
+    try {
+      const registro = await ItemCompraService.deleteItemCompra(id);
+      res.json(registro);
+    } catch (error) {
+      res.status(500).json({ error: error.message });
+    }
+  } else {
+    res.status(403).json({ error: 'Acesso negado' });
   }
 };
